added banned component. closed #39
