+++ pages/about.js
@@ -1,0 +1,14 @@
+/* @author Charles Liu */
+
+import Banner from '../components/shared/Banner'
+
+function about() {
+    return (
+        <div>
+          <Banner url="/images/iphone_background_landscape.png" title="ABOUT MSMR" />
+          <h1>Hello World</h1>  
+        </div>
+    )
+}
+
+export default about
